get git server working
