+++ tests/gitserver.go
@@ -1,0 +1,58 @@
+package tests
+
+import (
+	"bufio"
+	"bytes"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"net/http"
+	"net/textproto"
+	"os"
+	"os/exec"
+)
+
+func (run *runner) gitHandler(w http.ResponseWriter, r *http.Request) {
+	t := run.T
+	defer func() {
+		io.Copy(ioutil.Discard, r.Body)
+		r.Body.Close()
+	}()
+
+	cmd := exec.Command("git", "http-backend")
+	cmd.Env = []string{
+		fmt.Sprintf("GIT_PROJECT_ROOT=%s", run.dir),
+		fmt.Sprintf("GIT_HTTP_EXPORT_ALL="),
+		fmt.Sprintf("PATH_INFO=%s", r.URL.Path),
+		fmt.Sprintf("QUERY_STRING=%s", r.URL.RawQuery),
+		fmt.Sprintf("REQUEST_METHOD=%s", r.Method),
+		fmt.Sprintf("CONTENT_TYPE=%s", r.Header.Get("Content-Type")),
+	}
+
+	buffer := &bytes.Buffer{}
+	cmd.Stdin = r.Body
+	cmd.Stdout = buffer
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Run(); err != nil {
+		t.Fatal(err)
+	}
+
+	text := textproto.NewReader(bufio.NewReader(buffer))
+
+	code, _, _ := text.ReadCodeLine(-1)
+
+	if code != 0 {
+		w.WriteHeader(code)
+	}
+
+	headers, _ := text.ReadMIMEHeader()
+	head := w.Header()
+	for key, values := range headers {
+		for _, value := range values {
+			head.Add(key, value)
+		}
+	}
+
+	io.Copy(w, text.R)
+}
