+++ tests/lfsserver.go
@@ -1,0 +1,9 @@
+package tests
+
+import (
+	"net/http"
+)
+
+func (run *runner) lfsHandler(w http.ResponseWriter, r *http.Request) {
+	run.Fatalf("wat")
+}
